GRD: use 2021.3.1 EAP to build plugin
Split code using trusted project API for 212 and 213 platforms because of API changes in 2021.3.1
