+++ gradle-213.properties
@@ -1,16 +1,16 @@
 # Existent IDE versions can be found in the following repos:
 # https://www.jetbrains.com/intellij-repository/releases/
 # https://www.jetbrains.com/intellij-repository/snapshots/
-ideaVersion=IU-2021.3
-clionVersion=CL-2021.3
+ideaVersion=IU-213.6461-EAP-CANDIDATE-SNAPSHOT
+clionVersion=CL-213.6461-EAP-CANDIDATE-SNAPSHOT
 
 # https://plugins.jetbrains.com/plugin/8195-toml/versions
 tomlPluginVersion=213.5744.224
 # https://plugins.jetbrains.com/plugin/12775-native-debugging-support/versions
-nativeDebugPluginVersion=213.5744.186
+nativeDebugPluginVersion=213.6461.6
 # https://plugins.jetbrains.com/plugin/227-psiviewer/versions
 psiViewerPluginVersion=213-SNAPSHOT
 
 # please see https://www.jetbrains.org/intellij/sdk/docs/basics/getting_started/build_number_ranges.html for description
-sinceBuild=213.5744.223
+sinceBuild=213.6461
 untilBuild=221.*
