+++ src/212/main/resources/META-INF/platform-rust-core.xml
@@ -1,0 +1,10 @@
+<idea-plugin>
+    <projectListeners>
+        <listener class="org.rust.ide.security.RsTrustChangeNotifier"
+                  topic="com.intellij.ide.impl.TrustChangeNotifier"/>
+    </projectListeners>
+
+    <extensions defaultExtensionNs="com.intellij">
+        <editorNotificationProvider implementation="org.rust.ide.notifications.RsUntrustedNotificationProvider"/>
+    </extensions>
+</idea-plugin>
