+++ src/213/main/resources/META-INF/platform-rust-core.xml
@@ -1,0 +1,2 @@
+<idea-plugin>
+</idea-plugin>
