+++ src/main/resources/META-INF/rust-core.xml
@@ -3,6 +3,7 @@
 
     <depends>com.intellij.modules.lang</depends>
 
+    <xi:include href="/META-INF/platform-rust-core.xml" xpointer="xpointer(/idea-plugin/*)"/>
     <xi:include href="/META-INF/experiments.xml" xpointer="xpointer(/idea-plugin/*)"/>
 
     <extensions defaultExtensionNs="com.intellij">
@@ -1047,7 +1048,6 @@
 
         <editorNotificationProvider implementation="org.rust.ide.notifications.MissingToolchainNotificationProvider"/>
         <editorNotificationProvider implementation="org.rust.ide.notifications.NoCargoProjectNotificationProvider"/>
-        <editorNotificationProvider implementation="org.rust.ide.notifications.RsUntrustedNotificationProvider"/>
 
         <notificationGroup id="Rust Plugin" displayType="BALLOON"/>
         <notificationGroup id="Rust Build Log" displayType="NONE"/>
@@ -1231,8 +1231,6 @@
     <projectListeners>
         <listener class="org.rust.lang.core.macros.MacroExpansionManagerWaker"
                   topic="org.rust.cargo.project.model.CargoProjectsService$CargoProjectsListener" />
-        <listener class="org.rust.ide.security.RsTrustChangeNotifier"
-                  topic="com.intellij.ide.impl.TrustChangeNotifier"/>
     </projectListeners>
 
     <actions>
